refactor: Remove ProfileEdit Component
